requirements.txt
langchain>=0.0.352
python-dotenv>=1.0.0
openai>=1.3.0
solana>=0.30.2
anchorpy>=0.18.0
borsh-construct>=0.1.0
construct-typing>=0.6.2 